+++ src/main/java/net/brdle/collectorsreap/common/item/food/Nutrition.java
@@ -130,7 +130,8 @@ public class Nutrition {
 		.nutrition(14).saturationMod(0.75F)
 		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 3600, 0), 1F).build();
 	public static final FoodProperties SALMON_TARTARE = (new FoodProperties.Builder())
-		.nutrition(6).saturationMod(0.65F)
+		.nutrition(7).saturationMod(0.65F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 0), 1F)
 		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 3600, 0), 1F).build();
 	public static final FoodProperties POTATO_FRITTERS = (new FoodProperties.Builder())
 		.nutrition(7).saturationMod(1F)
@@ -200,42 +201,37 @@ public class Nutrition {
 		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 1200, 0), 1F)
 		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 3600, 0), 1F).build();
 	public static final FoodProperties SPICY_GRENADINE_JELLY = (new FoodProperties.Builder())
-		.nutrition(6).saturationMod(2.35F)
-		.effect(() -> new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 1800, 0), 1F)
+		.nutrition(7).saturationMod(1.2F)
+		.effect(() -> new MobEffectInstance(MNDCompat.getBeneficialPungent(), 1200, 0), 1F)
+		.effect(() -> new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 1200, 0), 1F)
 		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 1200, 0), 1F).build();
 	public static final FoodProperties POMEGRANATE_SMOOTHIE = (new FoodProperties.Builder())
 		.nutrition(3).saturationMod(0.2F)
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 600, 0), 1F)
-		.effect(() -> new MobEffectInstance(NeapolitanCompat.getAgility(), 600, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(NeapolitanCompat.getAgility(), 600, 0), 1F).build();
 	public static final FoodProperties LIMEADE = (new FoodProperties.Builder())
 		.alwaysEat()
 		.nutrition(3).saturationMod(0.2F)
-		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F).build();
 	public static final FoodProperties STRONG_LIMEADE = (new FoodProperties.Builder())
 		.alwaysEat()
 		.nutrition(3).saturationMod(0.2F)
-		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 1), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 1), 1F).build();
 	public static final FoodProperties BERRY_LIMEADE = (new FoodProperties.Builder())
 		.alwaysEat()
 		.nutrition(5).saturationMod(0.2F)
-		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F).build();
 	public static final FoodProperties PINK_LIMEADE = (new FoodProperties.Builder())
 		.alwaysEat()
 		.nutrition(5).saturationMod(1F)
 		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 1200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 1200, 0), 1F).build();
 	public static final FoodProperties MINT_LIMEADE = (new FoodProperties.Builder())
 		.alwaysEat()
 		.nutrition(6).saturationMod(0.35F)
 		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F)
-		.effect(() -> new MobEffectInstance(NeapolitanCompat.getBerserking(), 1200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(NeapolitanCompat.getBerserking(), 1200, 0), 1F).build();
 	public static final FoodProperties DRAGONS_PASSION = (new FoodProperties.Builder())
 		.alwaysEat()
 		.nutrition(5).saturationMod(1F)
@@ -247,65 +243,91 @@ public class Nutrition {
 			false,
 			true
 		), 1F)
-		.effect(() -> new MobEffectInstance(CompatEffects.getSpitting(), 800, 1), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CompatEffects.getSpitting(), 800, 1), 1F).build();
 	public static final FoodProperties LIME_GREEN_TEA = (new FoodProperties.Builder())
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(CompatEffects.getVitality(), 3600, 0), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 1200, 0), 1F).build();
 	public static final FoodProperties POMEGRANATE_BLACK_TEA = (new FoodProperties.Builder())
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(CompatEffects.getMaturity(), 600, 0), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 1200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 1200, 0), 1F).build();
 	public static final FoodProperties VERNAL_PURGE = (new FoodProperties.Builder())
 		.alwaysEat()
-		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 0), 1F).build();
 	public static final FoodProperties STRONG_VERNAL_PURGE = (new FoodProperties.Builder())
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 200, 1), 1F)
-		.effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 200, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 200, 0), 1F).build();
 	public static final FoodProperties LIMBO_BREW = (new FoodProperties.Builder())
 		.alwaysEat()
-		.effect(() -> new MobEffectInstance(MobEffects.HUNGER, 600, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(MobEffects.HUNGER, 600, 0), 1F).build();
 	public static final FoodProperties STRONG_LIMBO_BREW = (new FoodProperties.Builder())
 		.alwaysEat()
-		.effect(() -> new MobEffectInstance(MobEffects.HUNGER, 600, 1), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(MobEffects.HUNGER, 600, 1), 1F).build();
 	public static final FoodProperties SWEET_RECOVERY = (new FoodProperties.Builder())
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(MobEffects.HEAL, 1, 1), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 600, 2), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 600, 2), 1F).build();
 	public static final FoodProperties LONG_SWEET_RECOVERY = (new FoodProperties.Builder())
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(MobEffects.HEAL, 1, 0), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 1200, 1), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 1200, 1), 1F).build();
 	public static final FoodProperties STRONG_SWEET_RECOVERY = (new FoodProperties.Builder())
 		.alwaysEat()
 		.effect(() -> new MobEffectInstance(MobEffects.HEAL, 1, 2), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 300, 3), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.REBOUND.get(), 300, 3), 1F).build();
 	public static final FoodProperties DELUXE_SALAD = (new FoodProperties.Builder())
 		.nutrition(10).saturationMod(0.75F)
 		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 0), 1F)
 		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 600, 0), 1F)
-		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 600, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 600, 0), 1F).build();
 	public static final FoodProperties CREAM_CHEESE = (new FoodProperties.Builder())
 		.nutrition(4).saturationMod(0.5F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 1), 1F).build();
+	public static final FoodProperties PRAWN_SPREAD = (new FoodProperties.Builder())
+		.nutrition(9).saturationMod(0.85F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 1), 1F).build();
+	public static final FoodProperties CARBONARA_PASTA = (new FoodProperties.Builder())
+		.nutrition(14).saturationMod(0.8F)
 		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 1), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 6000, 0), 1F).build();
+	public static final FoodProperties STUFFED_PASTA_SHELLS = (new FoodProperties.Builder())
+		.nutrition(12).saturationMod(0.9F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 1), 1F)
+		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 3600, 0), 1F).build();
+	public static final FoodProperties BULLET_PEPPER_POPPER = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(0.75F)
+		.effect(() -> new MobEffectInstance(MNDCompat.getBeneficialPungent(), 400, 0), 1F)
+		.effect(() -> new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 400, 0), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 1), 1F).build();
+	public static final FoodProperties CANNOLI = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(1.25F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F).build();
+	public static final FoodProperties COFFEE_CANNOLI = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(1.25F)
+		.effect(() -> new MobEffectInstance(CompatEffects.getCaffeinated(), 400, 2), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F).build();
+	public static final FoodProperties CHOCOLATE_CANNOLI = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(1.25F)
+		.effect(() -> new MobEffectInstance(NeapolitanCompat.getSugarRush(), 600, 1), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F).build();
+	public static final FoodProperties MINT_CANNOLI = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(1.25F)
+		.effect(() -> new MobEffectInstance(NeapolitanCompat.getBerserking(), 400, 0), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F).build();
+	public static final FoodProperties STRAWBERRY_CANNOLI = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(1.25F)
+		.effect(() -> new MobEffectInstance(MobEffects.HEAL, 1, 0), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F).build();
+	public static final FoodProperties VANILLA_CANNOLI = (new FoodProperties.Builder())
+		.nutrition(8).saturationMod(1.25F)
+		.effect(() -> new MobEffectInstance(NeapolitanCompat.getVanillaScent(), 200, 0), 1F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 1), 1F).build();
 	public static final FoodProperties STRAWBERRY_JAM_BUN = (new FoodProperties.Builder())
 		.nutrition(8).saturationMod(0.55F)
 		.effect(() -> new MobEffectInstance(MobEffects.HEAL, 1, 0), 1F)
-		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 0), 1F)
-		.build();
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 600, 0), 1F).build();
 	public static final FoodProperties POMEGRANATE_BEAN_SALAD = (new FoodProperties.Builder())
 		.nutrition(8).saturationMod(0.35F)
 		.effect(() -> new MobEffectInstance(CREffects.VOLATILITY.get(), 400, 0), 1F)
@@ -485,6 +507,7 @@ public class Nutrition {
 		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 3600, 0), 1F).build();
 	public static final FoodProperties CRAB_LASAGNA = (new FoodProperties.Builder())
 		.nutrition(10).saturationMod(0.85F)
+		.effect(() -> new MobEffectInstance(CREffects.CORROSION.get(), 400, 1), 1F)
 		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 6000, 0), 1F).build();
 	public static final FoodProperties BUTTERED_LEGS = (new FoodProperties.Builder())
 		.nutrition(10).saturationMod(0.9F).build();
@@ -492,9 +515,9 @@ public class Nutrition {
 		.nutrition(9).saturationMod(0.9F).build();
 	public static final FoodProperties LAND_AND_SEA_BURGER = (new FoodProperties.Builder())
 		.nutrition(12).saturationMod(0.8F)
-		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 6000, 0), 1F)
 		.effect(() -> new MobEffectInstance(ModEffects.COMFORT.get(), 3600, 0), 1F)
-		.effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 200, 0), 1F).build();
+		.effect(() -> new MobEffectInstance(MobEffects.REGENERATION, 200, 0), 1F)
+		.effect(() -> new MobEffectInstance(ModEffects.NOURISHMENT.get(), 6000, 0), 1F).build();
 	public static final FoodProperties CLAM_MEAT = (new FoodProperties.Builder())
 		.nutrition(3).saturationMod(0.3F).build();
 	public static final FoodProperties CLAM_ROLL = (new FoodProperties.Builder())
@@ -551,7 +574,7 @@ public class Nutrition {
 		new MobEffectInstance(NeapolitanCompat.getAgility(), 200, 0), () ->
 		new MobEffectInstance(MobEffects.JUMP, 200, 2)));
 	public static final FoodProperties VANILLA_GUMMY = GUMMY(() ->
-		new MobEffectInstance(NeapolitanCompat.getVanillaScent(), 200, 0));
+		new MobEffectInstance(NeapolitanCompat.getVanillaScent(), 400, 0));
 	public static final FoodProperties CHOCOLATE_GUMMY = GUMMY(() ->
 		new MobEffectInstance(NeapolitanCompat.getSugarRush(), 200, 4));
 	public static final FoodProperties MINT_GUMMY = GUMMY(List.of(() ->
@@ -584,7 +607,7 @@ public class Nutrition {
 	public static final FoodProperties BLACK_TEA_GUMMY = GUMMY(() ->
 		new MobEffectInstance(CompatEffects.getMaturity(), 200, 2));
 	public static final FoodProperties COFFEE_GUMMY = GUMMY(() ->
-		new MobEffectInstance(CompatEffects.getCaffeinated(), 400, 3));
+		new MobEffectInstance(CompatEffects.getCaffeinated(), 400, 2));
 	public static final FoodProperties PRICKLY_PEAR_GUMMY = GUMMY(List.of(() ->
 		new MobEffectInstance(CompatEffects.getSpite(), 1800, 0), () ->
 		new MobEffectInstance(MobEffects.REGENERATION, 400, 0)));